(Feat): Open docs in same page with back button instead of new [pv622]
